+++ geometric/coordinate_systems/internal_base.py
@@ -432,65 +432,3 @@ class InternalCoordinateSystemBase(ABC):
             # Figure out the further change needed
             dQ1 = dQ1 - dQ_actual
             xyz1 = xyz2.copy()
-
-
-class SimpleIC(InternalCoordinateSystemBase, ABC):
-    """
-    Simple internal coordinate system: internals are used directly
-    """
-
-    def __init__(self, molecule):
-        super(SimpleIC, self).__init__(molecule)
-
-        self.Internals = []
-
-    def add(self, dof):
-        if dof not in self.Internals:
-            self.Internals.append(dof)
-
-    def calculate(self, xyz):
-        answer = []
-        for Internal in self.Internals:
-            answer.append(Internal.value(xyz))
-        return np.array(answer)
-
-    def derivatives(self, xyz):
-        self.calculate(xyz)
-        answer = []
-        for Internal in self.Internals:
-            answer.append(Internal.derivative(xyz))
-        # This array has dimensions:
-        # 1) Number of internal coordinates
-        # 2) Number of atoms
-        # 3) 3
-        return np.array(answer)
-
-
-class MixIC(InternalCoordinateSystemBase, ABC):
-    """
-    Mixing internal coordinate systems: Internals are linear combination of other internals
-
-    The over-complete IC system is stored in self.Prims
-    """
-
-    def __init__(self, molecule):
-        super(MixIC, self).__init__(molecule)
-
-        self.Prims: SimpleIC = None
-
-    def add(self, dof):
-        # for now we are not adding anything here, that is done when we are building the Prims
-        pass
-
-    def calculate(self, coords):
-        """ Calculate the DLCs given the Cartesian coordinates. """
-        PrimVals = self.Prims.calculate(coords)
-        Answer = np.dot(PrimVals, self.Vecs)
-        # To obtain the primitive coordinates from the delocalized internal coordinates,
-        # simply multiply self.Vecs*Answer.T where Answer.T is the column vector of delocalized
-        # internal coordinates. That means the "c's" in Equation 23 of Schlegel's review paper
-        # are simply the rows of the Vecs matrix.
-        # print np.dot(np.array(self.Vecs[0,:]).flatten(), np.array(Answer).flatten())
-        # print PrimVals[0]
-        # raw_input()
-        return np.array(Answer).flatten()
